refactor(cu): move common worker side-effects into clients. worker -> worker/evaluator and refactor
